Allow references to be added to Class.Schemes.
Restored functionality to allow notes and references to
be added to saved subfields of classification schemes, in
addition to the top-level field.

This requires that a field get saved to the database first
before attempting to create the reference.  This restriction
may be lifted in the future but for now the feature is
functional again.

See #259

Change-Id: I242c2c2079ec05c6a7e6355972ea0c3d8ab93c9f
Reviewed-on: http://review.source.gogoego.com/2327
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>