ProxyField Update for ForiegnKey*PrimitiveField
ProxyField can now take a table name for a given
ForeignKeyPrimitiveFeld or ForeignKeyListPrimitiveField,
which is required in the database.  Sets it to "null" by
default to make it consistent with the client, but sets
it to a string value nonetheless, such that database
errors won't occur when using ProxyField on the server.

Change-Id: I524ab556a467e79e70eede9e73ff17b7cf0dcc55
Reviewed-on: http://review.source.gogoego.com/2333
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>