RedListSource emits warning, In-Place L/W.
Now the RedListSource, when most recent-needs
updating appears, will cause a warning to appear in
the migration report.  The data block will be
converted over normally, though.

For InPlaceLandWaterProtectedPercentProtected, since this
field is now a band, translation into a number is not
needed.  Set this to a string field.

See #401
See #404

Change-Id: I160c0a0384effaceb505b2c2d7068bfe6cd29e58
Reviewed-on: http://review.source.gogoego.com/2358
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>