DEMPanel locking fix: allow if, deny all.
Changed the logic such that the status check
ensures that assessments are either unlocked
OR the user has the lock when allowing editing
of assessments; otherwise, deny editing.

Also, changed behavior such that it will keep
you in whichever editing mode you select as you
move through assessments, as opposed to forcing
you into edit mode when you wish to be in read-
only mode.

Change-Id: Ie147297686a526eb71f25459da7d9ce3e1b38db4
Reviewed-on: http://review.source.gogoego.com/2369
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>