Correct Expert System result display.
Expert system results, for some reason, were
chopping the first letter of additional criteria
strings.  I altered the code so that it does not
do this anymore, but left a means of quickly
reverting that behavior if necessary, as I'm not
sure why it was originally being done in the first
place.

Also, added a quick test case to ensure this can
be checked in the future.

See #297

Change-Id: I364ae0b51131f0edca71f9d36ec5d6f727f52763
Reviewed-on: http://review.source.gogoego.com/2370
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>