Corrected cache invalidation process.
Was attempting to remove entries based on an integer
instead of a string, which caused a cache-miss, resulting
in stale results later when the cache was believed to
be cleared.

This affects the child listing on the taxon home page,
since creating new taxa would not properly evict the
cache for the parent taxon.

See #411

Change-Id: I9f9ee1428ad41b834e44fc635bd2e503e589b68a
Reviewed-on: http://review.source.gogoego.com/2376
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>