Database-backed file attachments. (!)
Implemented database-backed file attachment system that
supports file upload, relating a file to one or more
saved narrative fields for any given assessment, and
the ability to make changes to the metadata post-upload
for information such as publish status, file name, as
well as marking which fields an attachment is associated
with. Also, any file can viewed or downloaded via a
button in the interface.

Additionally, replacing an uploaded file is also supported.
A replacing will keep the same metadata properties but
will be marked as a new "version" -- the version information
is essentially the same as edit information for an assessment
or note, tracking the date and user who made the change. An
version is also created upon initial upload.

Finally, added a means in the UI to add attachments from the
DEM toolbar as well as from any saved narrative fields that
supports attachments.

This change requires new database tables, which can be
generated via Hiberate tools or by using the SQL below:

CREATE TABLE fieldattachment
(
  id integer NOT NULL,
  "name" character varying(1024) NOT NULL,
  "key" character varying(1024) NOT NULL,
  publish boolean NOT NULL,
  CONSTRAINT fieldattachment_pkey PRIMARY KEY (id)
)
WITH (
  OIDS=FALSE
);

CREATE TABLE fieldattachment_edit
(
  editid integer NOT NULL,
  fieldattachmentid integer NOT NULL,
  CONSTRAINT fk2f6c176c6d74d570 FOREIGN KEY (fieldattachmentid)
      REFERENCES fieldattachment (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION,
  CONSTRAINT fk2f6c176cb68b1954 FOREIGN KEY (editid)
      REFERENCES edit (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION
)
WITH (
  OIDS=FALSE
);

CREATE TABLE fieldattachment_field
(
  fieldid integer NOT NULL,
  fieldattachmentid integer NOT NULL,
  CONSTRAINT fkbe2723d8581ac28a FOREIGN KEY (fieldid)
      REFERENCES field (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION,
  CONSTRAINT fkbe2723d86d74d570 FOREIGN KEY (fieldattachmentid)
      REFERENCES fieldattachment (id) MATCH SIMPLE
      ON UPDATE NO ACTION ON DELETE NO ACTION
)
WITH (
  OIDS=FALSE
);

See #52
See #334

Change-Id: I9119f92c670bdcec98329fc5b282164d2f72ef65
Reviewed-on: http://review.source.gogoego.com/2389
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>