Fixed a FIXME to prevent duplicate taxa creation.
Check to ensure no taxa on the same level within the
same kingdom can be created with the same name.  This
check is performed at the TaxomaticIO API level.

Also, changed the call to this API for the Working Set
import feature in order to check against the taxon level
as well when searching for an existing taxa by name.

See #413

Change-Id: I9516fb59d1898f2355edd3fce1afc35a168c8918
Reviewed-on: http://review.source.gogoego.com/2396
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>