Window standardization.
Standardized all the Windows such that they either
extend BasicWindow or are created via WindowUtils.
This includes ensuring the window has a title and
icon, is modal by default, the buttons are aligned
to the center, and they have buttons to close and
maximize the window.  Specific implementations can
override these properties as necessary, but this
initial creation process should have all windows
and popup alerts have a standard look and feel, and
changes to the standard can be made easily.

Also, did some code cleanup and removed an unused
class.

Change-Id: I2cde7ab2b521cdb0f590eedb503a182befd0cd33
Reviewed-on: http://review.source.gogoego.com/2399
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>