Gear menu's adminstrative tools cleanup.
The admin tools panel now lazy-loads its contents,
instead of loading all of them at once.  Also,
added some FIXMEs for the pages that aren't working
and need client and server-side improvements, such
as building a form on the client side, in order to
defeat permissions issues.

Also, set the NativeDocumentFactory's default instance
to the authenticated version of the NativeDocument
such that any requests using the NativeDocumentFactory
will be authenticated.  This is especially important
for the Integrity validation which is based on the
utility QueryBuilder, which uses the NativeDocumentFactory.

Change-Id: Ia187b16b9215c9a7f644b73fc7745825b5113ec7
Reviewed-on: http://review.source.gogoego.com/2400
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>