Update Commons DBCP and Pool to Viewmaster version
SIS-2 core was running with older versions of DBCP and Pool that
don't properly handle schemas. I am not sure what consequences
if any this update might have on core SIS, so please verify before
merging this change. But I don't think there should be any
negative side effects.

This is not an update to the latest versions of DBCP and Pool,
just the ones that were verified and tested in July 2010.

Change-Id: I153c1890425b923e5f670115692b14064b192ba7
Reviewed-on: http://review.source.gogoego.com/2427
Reviewed-by: Rob Heittman <rob.heittman@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>