Changed Kingdom listing to not be explicit.
When listing Kingdoms in SIS, it safely assumes
that it should list the four kingdoms.  However,
for installations where not all kingdoms are
present, this causes an exception.

Replaced the code to make it search for all the
active Kingdoms and return that set, instead of
using the four explicit kingdoms.  The original
ordering is still persisted here, so it will
look no different, but for installations with
only a kingdom or two, they can now use the
system as well.

Change-Id: I8c6fbf41cb1d0c2cc1d40f12f5a610c4966d3542
Reviewed-on: http://review.source.gogoego.com/2440
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>