Bug Fix: null values throw exception on display
Some types of lists, like SISBooleanList, throw NPEs
when a null value is passed, which is logical for a
new model.  Fixed the logic such that it checks for
nulls and returns a value as if it were passed in as
empty or not found.

This affect classification schemes because the viewer
was not picking up on this when trying to render the
cleaned version of data.  Instead, no friendly values
would show.

Change-Id: I3c27c27b6a5a007ac1e36bf52879d123da7b92eb
Reviewed-on: http://review.source.gogoego.com/2479
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>