Allow assessments to exist without a region.
Use/Trade assessents do not need a region.  Updated
key components of SIS that rely on the region field
being populated to no longer have this reliance. On
the server, proper checks are made.  On the client,
None or N/A are shown when no region is available.

Also, when creating new Use/Trade assessments, the
region and endemic fields are hidden, as they now
no longer need to be populated.

See #240

Change-Id: I27f9b6a15a001f171cc428087884d83824f976d7
Reviewed-on: http://review.source.gogoego.com/2481
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>