Process published assessments from cache.
Generate a cache file for published assessments
that state which assessments should be processed
when converting published assessments, instead of
reading through the file system.  This ensures no
ghost assessments come across in the conversion
step, only the assessments denoted in the Taxon
files.

This also allows for a means of published a subset
of assessments based on a taxon id, for testing
purposes, which means data runs can be done very
quickly for a small set of species.

Change-Id: Ie3d30dd335632bbd5806f4107f87aa633bc0a69e
Reviewed-on: http://review.source.gogoego.com/2494
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>