Only allow editing credit text for legacy data.
Changed the code to show the edit button only
when there is text being displayed in the textbox
based on legacy information.  Otherwise, the edit
button is hidden.

Once the manage credits window gets used, the text
portion will be nulled, thus the ability to edit
will be lost for good.

See #439

Change-Id: I13fb8ca5c3d2a92c55761820bd2bd1316eaa35b1
Reviewed-on: http://review.source.gogoego.com/2527
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>