Added the bad C2b assessment to ExpertSystem tests
As described in the ticket, includes only the population
size of 230 and extreme fluc for mature individuals set
to Yes.  This alone should not trigger C2b as C2 is not
yet met, but it does.

Also, added a good version that does include the C2 point
to ensure that it DOES come out as C2b when there is a
continuing decline.

See #437

Change-Id: I57282e230f8d3bb3384cea601ff1dd5bdcb51660
Reviewed-on: http://review.source.gogoego.com/2528
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>