Expert system logic correction for CR's C2b
This was an order of operations issue.  Re-ordered
the comparisons to ensure that C2a & C2b are taken
into account only if C2 is a factor, otherwise,
both are disregarded.

Test cases from previous commit now test green, and
the badC2b case now comes up as EN (D) due to the
population size, which is correct.

See #437

Change-Id: I91214182decb977258c399ffbf58bbd533dae4ec
Reviewed-on: http://review.source.gogoego.com/2529
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>