Expert System code cleanup.
Boiled down much of the expert system code to a single
subclass, since the criteria within the different categories
only vary, typically, by a compared number (i.e. a reduction
of 80% for CR A2 versus one of 50% for EN A2).  This ensures
consistency in the calculator and makes any future changes,
tweaks, or testing much easier.

Also, optimized some code to be faster and/or more readable.

A random commit, yes, but not in vain -- to accommodate #131,
more changes at this level will be needed, so this is a first
step in this direction.

See #131

Change-Id: Icfd43edf42ae643b54122e9fad317df7304c311a
Reviewed-on: http://review.source.gogoego.com/2536
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>