CriteriaLevel parse individual criterion to string
Implemented an tiered Criteria Level implementation
that can build a tree based on any number of individual
criterion (such as A1a or B1cii) and produce a full
criteria string.

This is the beginnings of being able to store individual
criteria met in the database and, more importantly, read
and present that information back.

The expert system should change to use this later to provide
a full range of information after it runs such that it can
be cleanly pushed to the database.

Also added test cases for this to ensure it doesn't break.

See #131

Change-Id: I0a9bdba004bbadc384f23847e14474d0645fb50c
Reviewed-on: http://review.source.gogoego.com/2538
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>