New & Improved Taxon Synonym Editor
New implementation of the taxon synonym editor,
based on Ext widgets and a cleaner UI style. Other
improvments include:

* Single synonym editing, for clarity
* Listing of all synonymns clearly visible
* Removed default "high level" text
* Prominent display of generated friendly name
* Cancel changes

Also, made a couple correcting to how Synonym
friendly names are generated -- they are now
created based on the taxon level, not just what
data is available.  Also, for Infraranks, null
will now be returned if a correct ID is not
supplied to the generation method, it will not
just assume subspecies or variety.

See #436

Change-Id: I2018bcdb5e8d51075ce3546da34768b906632248
Reviewed-on: http://review.source.gogoego.com/2539
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>