Expert System Code Cleanup, CriteriaSet, Tests
More code cleanup to the expert system to remove
duplicate and unnecessary code, and simplified
the criteria string resolution process by changing
the string building strategy -- filtering everything
through CriteriaSet.

The CriteriaSet stores a tree representation of all
criteria met for a particular category.  It also
holds a list containing the full name of each individual
criteria met, and can generate on-the-fly a full
string representation of a category for a criteria.

CriteriaSets can also be built from a field such that
only the appropriate matching criteria are included
in the set.  So, if D meets EN and E meets VU, making
a CriteriaSet for VU will show D;E, whereas making one
for EN will only show D.

This is big since we can now re-generate this information
from memory or, more strategically, a database.  Hence,
this affects #131.

Finally, added additional test case coverage for the
expert system and the criteria level generator, and
refactored any code affected by the expert system changes.

See #131

Change-Id: If94dd4e135db2d158c4c3189800af75bc07ebe59
Reviewed-on: http://review.source.gogoego.com/2540
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>