Username uniqueness is case insensitive.
Looking for a user by username should not be
case sensitive.  So SomeUser@example.com should
be considered the same as sOMEuSER@example.com.

This comes into play when evaluating whether a
user exists already when creating new users as
well as when search for users or modifying
existing users' usernames.

See #441

Change-Id: I1beaf931825a041d94a653612f46f2c689d4b85f
Reviewed-on: http://review.source.gogoego.com/2547
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>