Store Complete Expert System Results
Running the expert system will now store all
information about the results to the database
in additional fields in RedListFuzzyResults.

Added a proxy field to get and set information,
and updated the database scripts to add new
columns for this field.

Also, added a parser in CriteriaSet that can
parse an entire criteria string and build a
criteria set from it, along with test cases of
many different variations of criteria strings
to ensure this works properly.

See #131

Change-Id: I8dad0907d1d72ec2d6169e9e84376fb8f971a917
Reviewed-on: http://review.source.gogoego.com/2560
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>