Objective denotation of uncertainly in Bool.Range
Uncertainly within BooleanRange is now denoted more
objectively with the single character "U". When
selecting Unknown from a BooleanRange drop-down, it
will insert that character in the database, as opposed
to ".5".

Made changes to ensure that existing ".5" from SIS 1
are converted over to "U" for SIS 2.  Also, changed
the widget to put in "U" instead of ".5".

Finally, to be consistent with #57, changed the expert
system to check for "U" instead of ".5", and disregard
anything marked with a "U". Custom values of ".5" will
now be analyzed again.

See #57
See #430

Change-Id: I78ae4757c7b63cf3bc97b72aa2dac8ec535b1f8d
Reviewed-on: http://review.source.gogoego.com/2563
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>