Initial plumbing for a generic exporting tool.
Generic Exporter should be able to export to anything
that DBSession supports.  It will first find the source
database (SIS) and set it to the appropriate schema, based
on a choice made by the user, "access" for example.  Then,
this schema will be located in the source database.  Any
tables found in this schema that begin with "export_" will
be exported and assumed to have a column called
"assessmentid".  We can remove the naming restriction if
needed.

There is also plumbing to specify a template database, that
will be used to simply copy over metadata that needs to be
in the database, such as regions.  In the case of Access,
we may need to actually copy and use this template database
directly instead of copying it, but that can be adjusted as
needed.

See #85

Change-Id: I03c53dd8020fc7a762e501e6c4a1e36a41910b51
Reviewed-on: http://review.source.gogoego.com/2653
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>