RedListHistory field now tracks history text
New field RedListHistory tracks the information formerly
stored in RedListCriteria.rlHistoryText.

Updated the taxomatic to use this new field to store this
information, and changed the AssessmentConverter to put
this data in this new field and not in RedListCriteria.

TODO: create the field, update the existing RedListCriteria
field to not use this information.

See #462

Change-Id: I7a72d583099e2a58429d8a5bab8a05bcc932ef92
Reviewed-on: http://review.source.gogoego.com/2658
Tested-by: Rasanka Jayawardana <Rasanka.Jayawardana@iucn.org>
Reviewed-by: Rasanka Jayawardana <Rasanka.Jayawardana@iucn.org>