Corrected validation to check for nulls, authority
Validation rules were checking for blank, not null, which
allowed null to pass as a value instead of blank.  This
threw an uncaught exception in the browser, which was bypassed,
and allowed this bad data into the database.  Now, null and
blank values are treated the same, and the appropriate error
message will display.

Also, added support for managing authorities for levels Kingdom
through Genus, as well as removing the ability to change the
taxon level for Kingdom through Genus.

Finally, added authorities to the display rules such that they
will show when entered for Kingdom through Genus.

See #436
See #504

Change-Id: I97d07e5dce7c22f09094296f51cee4b109f782d5
Reviewed-on: http://review.source.gogoego.com/2677
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>