Smart sorting of classification scheme fields.
Fields that were classification schemes could vary
in their XML representations even though their data
was the same.  The current sorting method only orders
by field name which, of course, is the same for this
type of field.

Changed the logic such that if the field names are
the same, the compare logic must sort on the actual
data stored.  So, it explores the fields and primitive
fields, sorting on the raw values of the data.

See #511

Change-Id: I8c326e526e339365a68c1eb0b3f8077871147a44
Reviewed-on: http://review.source.gogoego.com/2694
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>