Allow table types, Use StringLiteral for CBoolean
DBSession now allows you to specify what table types
are eligible to be returned when creating a structure
based on the database metadata.  This means that not
just tables, but also views, can be returned in a
table listing.

Also, changed CBoolean's getLiteral function to return
the literal value 'true' or 'false' instead of the
numeric 1 or 0.  This is SIS-specific and should not
make it into any library merges, but we will need to
deal with this specifically for Postgresql later.

Change-Id: I219ac9e1f8912b481b8403c52e3414535db38970
Reviewed-on: http://review.source.gogoego.com/2695
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>