Generic Exporter improvements, export all tables
Further implementation of the generic exporter now allows for
any table to be exported that's in the schema.  Still enforces
that a table with assessment information should have an assessmentid
in the table, but all information is still pulled.

Also construct the taxon table based on the taxa in your working
set.

See #85

Change-Id: Ieb63c66521cd401d9da10c688587bef84b620728
Reviewed-on: http://review.source.gogoego.com/2697
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>