Allow user-entered name in support panel, API
The support window now allows the user to enter
their name as the reporter for a new ticket.  This
is pre-populated with their display name.

Also, updated the server to implement to the latest
support API for the SIS Assembla space, such that
tickets are now coming across with additional
supplied metadata.

See #474

Change-Id: Icab964bd5fe91d70f8f0a8f4a58672cfe5d77122
Reviewed-on: http://review.source.gogoego.com/2709
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>