Taxon Home Page Changes, Welcome TaxonomicNotes
Added the taxonomic notes field to the Taxon home page
with the ability to edit.  Also renamed "References"
to "Taxonomic Sources."

The TaxonomicNotes editor can also be accessed from the
Tools menu in the Assessment Data Browser.  Permissions
to use this tool are the same as the permissions to edit
the Common Names & Synonyms.

See #471

Change-Id: I14a401e9b96a6859056fd11b7a293db672945114
Reviewed-on: http://review.source.gogoego.com/2728
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>