Search for lookups for range primitive fields.
Lookup scan was only looking for foreign key
fields for lookups, but other fields may have
lookups as well (like range).  Added range to
the list of field types scanned.

For future reference, would be nice to have this
information about what a lookup table is for a
given primitive field in the lookups metadata
itself.

Change-Id: I19234d2d8d3737d7b4021d61c0f97d83c5734d75
Reviewed-on: http://review.source.gogoego.com/2738
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>