Classification Scheme read-only grid styling.
Styled the read-only classification scheme grids to have
a light border between the rows and cells, also moved the
view/edit and quick add buttons up to to the top of the
list and added some styling to make it more of a header.

For Research & ConservationActions, made the first column
double the width, since there's only one data point for
that Classification Scheme (the note).

See #454

Change-Id: Idad6ed0b79bcd34b190ad99e48d24e1c27536cdc
Reviewed-on: http://review.source.gogoego.com/2781
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>