Login process cleanup, AuthorizationCache.
Remove an unnecessary step in the login process
that involved hitting the server to grab a user's
permissions -- permissions come back with the
user's profile, so no need to do a second request.

Also, simplified AuthorizationCache to only have
the current user's permission groups cached (since
only one person is logged in to a browser) to
remove unneeded code and checks.

Centralized the NativeDocument instantiation to
only the AuthorizationCache.

Finally, added a toString() method to Permission
model that shows information on each permission;
should help with debugging.

Change-Id: I5e1260b4ae347a36ce36d385b64d483fda582408
Reviewed-on: http://review.source.gogoego.com/2793
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>