Database replication, generic exporters, more
Where do I begin...

Updated SISPersistentManager to allow for explicit and
dangerous creation of SessionFactories, as well as
the ability to run schema export on a given database
(with great power...).  Also, added a shutodwn method
to allow complete disconnect from a database.

Implemented a means of specifying a SessionFactory's
ID assignment pattern for all entities -- a backup or
target database, for example, will typically want all
IDs to be assigned, not generated, in any case.  This
is needed because of a known bug with Hibernate -- the
replicate function should handle this, but it doesn't.

Changed HBM configuration to cascade replicate where
appropriate.  This is needed for properly using the
replicate functionality in Hibernate.

Finally, code cleanup on the GenericExporter on up, and
added a middle layer of DatabaseExport as well.  Atop
this is the new working set exporter, which can create
(and zip when needed) a working set from a source into
a Postgres or H2 target database (if H2, zip it).

See #357

Change-Id: I2ea525b4b26ac91053916ab1489aa8658f9e03a1
Reviewed-on: http://review.source.gogoego.com/2797
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>