Implemented TabPanel for Admin. Tools as Menu
Changed the implementation of the administrative tools
section (in the gear menu) from being a TabPanel with
all tools to breaking down each option in a menu and
displaying a new window for each selection.

This provides more on-screen room, and also affords us
the option to make this into it's own page rather than
a popup.  Would like to see that happen down the line
to take advantage of the screen real estate for some of
the more cramped tools.

Also, added a new utility method in AuthorizationCache
to check if a feature can be used.

Change-Id: I847ba8b08361975ee963b1d3cacc00da55bb8a7d
Reviewed-on: http://review.source.gogoego.com/2823
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>