Expert System C2aii correction, code cleanup
Ensure that the subpopulation numbers are being
properly calculated and divided (that last step
was missing).  Also, added a few test cases for
this.

Cleaned up code for this function to make it more
obvious what each variable was used for and also
for function clarity when joining information
(which also pointed out an issue). Also, added
code comments to CR & EN for division information.

ExpertSystem tests green still with this commit.

Ready to Deploy #503

Change-Id: I7c9d00a3bd430c6a8e0d8418569fec7440ce503e
Reviewed-on: http://review.source.gogoego.com/2826
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>