Resize, reword ClassificationSchemeRowEditorWindow
Resized the window to be a bit shorter, as the additional
screen real estate was unnecessary here and in fact was
going below the fold.  Standardized this new height (and
the existing width) to be consistent across all editing
and viewing windows.

Also, reworded the select text for quick add to simply
say "Select," insuring a consistent width across each
window and keeping the drop-down control on-screen.

Ready to Deploy #538
Ready to Deploy #539

Change-Id: I13968cc654bebe21e284285ce4ea873779293e34
Reviewed-on: http://review.source.gogoego.com/2837
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>