Test Suite Additions: BasicHibernateTest
Part of the foundation of the SIS Test Suite, the
BasicHibernateTest is an extension of the BasicTest,
which always gives you test-mode debugging.  This
extension now starts up Hibernate and sets up your
database and Session based on your local_config
properties.

This functionality can be utilized by simply creating
a test class and having it extend this one.

To support this, made some changes to Debugger, to
allow it to ignore being set unless forced, and also
to catch and ignore an OSGi-specific exception, as
test cases are typically performed in a non-OSGi
environment.

Change-Id: I557354b8eef26536e9c35a0fa2fb399ce3a651ae
Reviewed-on: http://review.source.gogoego.com/2838
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>