Hibernate Lock Repository fixes, tests, inclusion.
The FileLocker now uses the Hibernate Lock Repository
implemenation instead of the PersistentLockRepository.

Bugs that were causing the Hibernate version to fail
have all been fixed in this commit, mainly due to
session objects being closed prematurely.  Also, the
HBM file configuration was a tad off.

To prevent regressions, added LockTesting to the test
suite, based on the new BasicHiberateTest, which
exercises the core functionality needed for locking.

This affects the offline export mechanism since there
is an option to lock assessments as "checked out" when
downloading, so this functionality is necessary. While
the PersistentLock repository works today, the code is,
by nature, not as maintainable as the Hibernate version,
since it uses direct SQL queries.

See #357
See #541

Change-Id: Ia5f88aa426aedfdd86fd2a4316206ebcfc0aac96
Reviewed-on: http://review.source.gogoego.com/2839
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>