Ensure 3-minute expiring save locks applied.
When a user saves changes to an assessment, that
assessment has a persistent save lock applied,
meaning that only that user can make changes to
that assessment.  After three minutes of inactivity,
the lock is automatically discarded.  The lock can
also be forcibly removed via the Manage Locks
administrative panel.

Ready to Deploy #32

Change-Id: I0336efb8afb8321ff4b1d3ab93d6290bbea00e6c
Reviewed-on: http://review.source.gogoego.com/2843
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>