Display selected taxon/virus for Named threats
For named threats, show the actual name of the
selected taxon or virus for each of the selected
rows.

Also, removed explicit checked for 8.5.2 (named virus)
and replaced with a check for hasVirus via the
ThreatViewerFactory, and ensure that "Unspecified"
prefix for taxon level genus and above displays in
all places where invasive taxa are displayed when
editing Threats.

Ready to Deploy #507

Change-Id: I6e9ee4e5e7257aac3048c15532d2b5786c91dbfa
Reviewed-on: http://review.source.gogoego.com/2875
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>