Remove region "2" (ID 13)
Added code to the SQL script that delete other duplicate
regions that will delete the region ID 13 altogether and
also set any assessments with this region to a deleted
state (they will appear in the trash bin).  Since these
assessments will be deleted in SIS 1, this probably will
not be an issue.

Ready to Deploy #543

Change-Id: I434a96c4c48eb18c8f131c58d9875b27ca66582f
Reviewed-on: http://review.source.gogoego.com/2883
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>