Apply, update count based on contextual ref group.
The count associated with a reference search is now
context-sensitive based on the group of references
you are attempting to search on.  For example, if you
are adding a reference to a taxon, the count number
will represent the number of *taxa* each reference is
used on.  If you are searching on a field, then the
count will represent the number of *fields* each
reference is used on.

This means that now the counting functionality is
available at all levels of reference management, as
opposed to only fields.

Additionally, the UI will now increment & decrement
the count when adding & removing references from a
source, respectively.  This count is completely a
client-side ordeal and may not be an accurate
reflection based on changes other users may have
made; a complete re-fetch of the search would need
to be done, which is costly, so this will provide
enough to make users smile.

A note here -- adding references at the assessment
level is NOT the same as adding references to a
field in that assessment.  Hence, the counts will
differ between assessment references (global) and
field references.

Finally, this commit also removed a couple unused
classes that implemented Referenceable; instead of
updating these unnecessary classes, they simply
get deleted here.

Ready to Deploy #453

Change-Id: I1f440a6286d615b3d0eac1d4b3b8dd15a4d718ca
Reviewed-on: http://review.source.gogoego.com/2902
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>