Portable Permissions oracle, update test cases
Create a portable permission oracle (can be used on
both the client and the server) which take an argument
for a PermissionDataSource.  This provides access to
taxa and working sets as needed for certain permitted
operations.

Also, updated the client to use this as the basis for
its implementation, and removed the one-off testing
implementation (SharedPermissionUtils).  The test cases
now use the same code as the actual implementation.

Finally, changed PermissionTesting from a script to be
a BasicHibernate test, added appropriate before and
after methods, and changed the test methods to make
interesting assertions as appropriate.

Change-Id: I715fb2b4786c586b4027a550c1e739270af53751
Reviewed-on: http://review.source.gogoego.com/2912
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Rasanka Jayawardana <Rasanka.Jayawardana@iucn.org>
Reviewed-by: Rasanka Jayawardana <Rasanka.Jayawardana@iucn.org>