NewAssessmentPanel UI & permissions cleanup
Cleaned up the UI a bit to make it more scalable, neat,
and have a similar look and feel to other SIS 2 panels.

Also, the panel will no longer display an empty list of
assessment types.  If you do not have permission, the
window simply will not open; a prompt will be displayed
instead.

See #564

Change-Id: I778684b3a5b3298dd64d4018ddfaa1fbc4fc5176
Reviewed-on: http://review.source.gogoego.com/2922
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>