Account for hidden asms when building vw_published
When creating the vw_published.vw_filter, account for
hidden assessment (where RedListHidden.value is true).
In practice, since the database does not store false
values, it is assumed that if there is a row here,
the row should be hidden.

Also, added RedListHidden to canonical names.

Ready to Deploy #562

Change-Id: I5933184ec543b7a17834c569225ed6a0b2b6e07a
Reviewed-on: http://review.source.gogoego.com/2919
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>