Intro. MultiDocumentRestlet, cache init change
MultiDocumentRestlet can take many request URIs in a
single POST, dispatch, and return the results in a
single request.  RESTful? No. Efficient? You bet.

For this, the caches that need to load when logging
in to SIS (Bookmarks, Definitions, Schemas, Regions,
etc), which previously gathered data each in a single
request, now gather this information in one request.
The client then dispatches this information to the
appropriate handler.  No differences should be
noticed here besides speed.

Also, committed a refactoring updated needed in the
now-defunct workflow application; a change in the
integrity API broke this.

Change-Id: If3c9a66f335cde7f9fdc4de0af40cbf95466db21
Reviewed-on: http://review.source.gogoego.com/2970
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>