Corrected B string generation for sub-criteria.
For the B string, the a subcriteria for B1 and
B2 was only being appended if A was matched for
severely fragmented, but was ignoring the
locations number data.  Updated the string
generation logic to include this information.

Also, added a test case for this particular
issue to prevent regression.  All existing test
cases still pass green with this change.

Ready to Deploy #568

Change-Id: Id313b7c4ce6a9c901dd930ff3e29faa0d8a20852
Reviewed-on: http://review.source.gogoego.com/2987
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>